Editor window updates with currently selected post
This uses another Fetch, with the premise that I'm only retrieving a
post preview in the sidebar. However, this isn't really the case,
and it's probably not necessary (if there is that much blog content, I
should be paginating anyway).

Because the extra fetch has deleterious impact on responsiveness, in
future, might lift posts state up & just have editor retrieve from
there.